Adds backend build and push workflow
Implements a new GitHub Actions workflow to build and push the backend Docker image to Docker Hub.

This workflow automates the process of building the backend application and pushing it to Docker Hub upon pull requests to the `dev` branch. It also defines various tags for the Docker image based on schedule, branch, pull request, semantic versioning, and SHA.

The Dockerfile uses a multi-stage build to create a smaller final image.