Adds backend v2 build and push workflow
Adds a new GitHub Actions workflow to build and push the backend v2 application.

The workflow is triggered on pull requests to the `feature/api/create-api-v2-with-java` branch,
when changes are made to the `api-v2/smart-elearning` directory or the workflow file itself.

It sets up JDK 17, builds a Docker image, logs into Docker Hub, and pushes the image.

Also disables the context load test for v2.